requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
torch
torchvision
pillow
shapely
mlflow
